+++ meetup2apricot/meetup_api.py
@@ -2,9 +2,10 @@
 
 from .http_response_error import MeetupApiError
 from .throttle import make_throttle
+from http import HTTPStatus
 
 
-class MeetupEventsRetriever:
+class MeetupApi:
 
     """Retrieves events and other data from Meetup via their API."""
 
@@ -34,6 +35,20 @@ class MeetupEventsRetriever:
         params = self.request_params()
         params.update(kwargs)
         response = self.session.get(url, params=params)
+        MeetupApiError.check_response_status(response)
+        return response.json()
+
+    def retrieve_event_json(self, meetup_id):
+        """Retrieve the JSON event by its Meetup ID."""
+        self.throttle.throttle()
+        url = self.build_url(self.group_url_name, "events", meetup_id)
+        response = self.session.get(url)
+        if response.status_code in [
+            HTTPStatus.FORBIDDEN,
+            HTTPStatus.NOT_FOUND,
+            HTTPStatus.GONE,
+        ]:
+            return None
         MeetupApiError.check_response_status(response)
         return response.json()